Refactored the session pallet
