+++ runtime/src/configs/mod.rs
@@ -5,3 +5,4 @@ pub mod transaction_payment;
 pub mod membership;
 pub mod im_online;
 pub mod offences;
+pub mod session;
